requirements.txt
immudb-py==0.9.2
Twisted==21.2.0
